print natural numbers upto or downto 98

+++ 0x02-functions_nested_loops/11-print_to_98.c
@@ -24,7 +24,7 @@ void print_to_98(int n)
 			_putchar('0' + x);
 			return;
 		}
-		_putchar('0' + x);
+		_putchar('0' + (int)x);
 		_putchar(',');
 		_putchar(' ');
 		x--;
